+++ JAM-AGOSTO/Assets/Assets/NoTransitable.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: da21526d58f2b7c4e8aafe80fbb29d9a, type: 3}
+  m_Name: NoTransitable
+  m_EditorClassIdentifier: 
+  tiles:
+  - {fileID: 11400000, guid: 2b9dca8d7e127244294735c8e5a79936, type: 2}
+  transitable: 0
