+++ JAM-AGOSTO/Assets/Assets/NoTransitable.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2c80e5e5341f15040b3c33e51fef9311
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
