+++ JAM-AGOSTO/Assets/Assets/Transitable.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: da21526d58f2b7c4e8aafe80fbb29d9a, type: 3}
+  m_Name: Transitable
+  m_EditorClassIdentifier: 
+  tiles:
+  - {fileID: 11400000, guid: 2fa99e96ef39d334282b7362767a46ea, type: 2}
+  transitable: 1
