+++ JAM-AGOSTO/Assets/Assets/Transitable.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc1476ff8e0ae3740a3d418b11ef76f3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
