+++ JAM-AGOSTO/Assets/ScriptableObjects/TileData.cs
@@ -3,17 +3,11 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Tilemaps;
 
+[CreateAssetMenu]
+
 public class TileData : ScriptableObject
 {
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
+    public TileBase[] tiles;
 
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
+    public float transitable;
 }
